Rename legacy.js to ios.js
